Django/project02/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.3.0
click==8.1.3
Django==4.0.5
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
